poly-edge-bot/requirements.txt
# ============================================================
# poly-edge-bot — Dependencies
# ============================================================
py-clob-client>=0.6.0
httpx>=0.27.0
pydantic>=2.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
numpy>=1.26.0
click>=8.1.0
rich>=13.7.0
aiohttp>=3.9.0
python-telegram-bot>=21.0
